third_party/blink/web_tests/accessibility/scroll-to-make-visible-div-overflow-expected.txt

Tests that scrolling to make an element visible successfully scrolls an arbitrary HTML element that has CSS overflow set to 'scroll'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Upper Target
5000-pixel box
Lower Target
PASS container.scrollTop is 0
PASS container.scrollTop >= minYOffset is true
PASS container.scrollTop <= maxYOffset is true
PASS container.scrollTop >= minYOffset is true
PASS container.scrollTop <= maxYOffset is true
PASS container.scrollTop >= minYOffset is true
PASS container.scrollTop <= maxYOffset is true
PASS successfullyParsed is true

TEST COMPLETE
